Revert "fix: changes limit from int32 to long (64-bit) to allow larger files to be uploaded"
This reverts commit c883a455088207aae0c96910b6c1e610ee5c40f2.

+++ src/Umbraco.Cms.Api.Management/ViewModels/TemporaryFile/TemporaryFileConfigurationResponseModel.cs
@@ -8,5 +8,5 @@ public class TemporaryFileConfigurationResponseModel
 
     public string[] AllowedUploadedFileExtensions { get; set; } = Array.Empty<string>();
 
-    public long? MaxFileSize { get; set; }
+    public int? MaxFileSize { get; set; }
 }
